+++ R/posest.R
@@ -25,20 +25,22 @@ create_df_posest <- function() {
 #'   \item{Contents}{Text content of product}
 #' }
 #' @param link URL to storm's archive page.
-#' @param msg Show link being worked. Default, FALSE.
 #' @seealso \code{\link{get_storms}}, \code{\link{posest}}
 #' @export
-get_posest <- function(link, msg = FALSE) {
+get_posest <- function(link) {
 
     # Check status of link(s)
     valid.link <- sapply(link, status)
     valid.link <- na.omit(valid.link)
-    if (length(valid.link) == 0)
-        stop("No valid links.")
 
     products <- purrr::map(valid.link, get_products) %>% purrr::flatten_chr()
 
-    products.posest <- purrr::map(filter_posest(products), posest)
+    products <- filter_posest(products)
+
+    # Set progress bar
+    p <- dplyr::progress_estimated(n = length(products))
+
+    products.posest <- purrr::map(products, posest, p)
 
     posest <- purrr::map_df(products.posest, dplyr::bind_rows)
 
@@ -51,13 +53,18 @@ get_posest <- function(link, msg = FALSE) {
 #' @details Given a direct link to a position estimate product, parse and return
 #' dataframe of values.
 #' @param link URL of a specific position estimate product
-#' @param msg Display each link as being worked; default is FALSE.
+#' @param p dplyr::progress_estimate.
 #' @return Dataframe
 #' @seealso \code{\link{get_posest}}
 #' @keywords internal
-posest <- function(link, msg = FALSE) {
+posest <- function(link, p) {
 
-    contents <- scrape_contents(link, msg = msg)
+    p$pause(0.5)$tick()$print()
+
+    contents <- scrape_contents(link)
+
+    # Replace all carriage returns with empty string.
+    contents <- stringr::str_replace_all(contents, "\r", "")
 
     # Make sure this is a public advisory product
     if (!any(stringr::str_count(contents,
@@ -70,6 +77,10 @@ posest <- function(link, msg = FALSE) {
     name <- scrape_header(contents, ret = "name")
     adv <- scrape_header(contents, ret = "adv")
     date <- scrape_header(contents, ret = "date")
+
+    if (getOption("rrricanes.working_msg"))
+        message(sprintf("Working %s %s Position Estimate #%s (%s)",
+                        status, name, adv, date))
 
     df <- df %>%
         tibble::add_row("Status" = status,