+++ R/prblty.R
@@ -9,20 +9,25 @@
 #'   \item{Contents}{Text content of product}
 #' }
 #' @param link URL to storm's archive page.
-#' @param msg Show link being worked. Default, FALSE.
 #' @seealso \code{\link{get_storms}}, \code{\link{prblty}}
 #' @export
-get_prblty <- function(link, msg = FALSE) {
+get_prblty <- function(link) {
 
     # Check status of link(s)
     valid.link <- sapply(link, status)
     valid.link <- na.omit(valid.link)
-    if (length(valid.link) == 0)
-        stop("No valid links.")
 
+    # Get all products for the current storm
     products <- purrr::map(valid.link, get_products) %>% purrr::flatten_chr()
 
-    products.prblty <- purrr::map(filter_prblty(products), prblty)
+    # Filter out prblty products
+    products <- filter_prblty(products)
+
+    # Set progress bar
+    p <- dplyr::progress_estimated(n = length(products))
+
+    # Work products
+    products.prblty <- purrr::map(products, prblty, p)
 
     prblty <- purrr::map_df(products.prblty, dplyr::bind_rows)
 
@@ -34,13 +39,18 @@ get_prblty <- function(link, msg = FALSE) {
 #' @details Given a direct link to a strike probability advisory product, parse
 #' and return dataframe of values.
 #' @param link Link to a storm's specific strike probability advisory product.
-#' @param msg Display each link as being worked; default is FALSE
+#' @param p dplyr::progress_estimate.
 #' @return Dataframe
 #' @seealso \code{\link{get_prblty}}
 #' @keywords internal
-prblty <- function(link, msg = FALSE) {
+prblty <- function(link, p) {
 
-    contents <- scrape_contents(link, msg = msg)
+    p$pause(0.5)$tick()$print()
+
+    contents <- scrape_contents(link)
+
+    # Replace all carriage returns with empty string.
+    contents <- stringr::str_replace_all(contents, "\r", "")
 
     # Make sure this is a strike probability product
     if (!any(stringr::str_count(contents, c("MIASPFAT", "MIASPFEP", "SPFAT", "MIAWRKSP"))))
@@ -50,6 +60,10 @@ prblty <- function(link, msg = FALSE) {
     name <- scrape_header(contents, ret = "name")
     adv <- scrape_header(contents, ret = "adv")
     date <- scrape_header(contents, ret = "date")
+
+    if (getOption("rrricanes.working_msg"))
+        message(sprintf("Working %s %s Strike Probability #%s (%s)",
+                        status, name, adv, date))
 
     # 15.0N  43.4W      43  1  X  X 44   16.8N  48.2W       X  4 16  2 22
     # 15.8N  45.9W       1 26  1  X 28