+++ R/tracking_chart.R
@@ -1,8 +1,26 @@
 #' @title al_tracking_chart
 #' @description Build tracking chart centered on Atlantic Basin.
-#' @param ... Additional parameters for ggplot2
+#' @param ... Additional parameters for \link{tracking_chart} and ggplot2
 #' @seealso \code{\link{tracking_chart}}
 #' @return ggplot2 object centered on Atlantic basin.
+#' @examples
+#' \dontrun{
+#' # Build map with white land areas, thin black borders
+#' al_tracking_chart(color = "black", size = 0.1, fill = "white")
+#'
+#' # 50nm resolution, no states
+#' al_tracking_chart(res = 50, states = FALSE, color = "black", size = 0.1,
+#'                   fill = "white")
+#'
+#' # 50nm resolution, coastlines only
+#' al_tracking_chart(countries = FALSE, res = 50, color = "black", size = 0.1,
+#'                   fill = "white")
+#'
+#' # Adding and modifying with ggplot functions
+#' al_tracking_chart(color = "black", size = 0.1, fill = "white") +
+#'     ggplot2::labs(x = "Lon", y = "Lat",
+#'     title = "Base Atlantic Tracking Chart")
+#' }
 #' @export
 al_tracking_chart <- function(...) {
     p <- tracking_chart(...)
@@ -15,6 +33,24 @@ al_tracking_chart <- function(...) {
 #' @param ... Additional parameters for ggplot2
 #' @seealso \code{\link{tracking_chart}}
 #' @return ggplot2 object centered on northeast Pacific basin.
+#' @examples
+#' \dontrun{
+#' # Build map with white land areas, thin black borders
+#' ep_tracking_chart(color = "black", size = 0.1, fill = "white")
+#'
+#' # 50nm resolution, no states
+#' ep_tracking_chart(res = 50, states = FALSE, color = "black", size = 0.1,
+#'                   fill = "white")
+#'
+#' # 50nm resolution, coastlines only
+#' ep_tracking_chart(countries = FALSE, res = 50, color = "black", size = 0.1,
+#'                   fill = "white")
+#'
+#' # Adding and modifying with ggplot functions
+#' ep_tracking_chart(color = "black", size = 0.1, fill = "white") +
+#'     ggplot2::labs(x = "Lon", y = "Lat",
+#'     title = "Base East Pacific Tracking Chart")
+#' }
 #' @export
 ep_tracking_chart <- function(...) {
     p <- tracking_chart(...)
@@ -34,6 +70,23 @@ ep_tracking_chart <- function(...) {
 #' @return Returns ggplot2 object that can be printed directly or have new
 #'     layers added.
 #' @seealso \code{\link[ggplot2]{aes}}
+#' @examples
+#' \dontrun{
+#' # Build map with white land areas, thin black borders
+#' tracking_chart(color = "black", size = 0.1, fill = "white")
+#'
+#' # 50nm resolution, no states
+#' tracking_chart(res = 50, states = FALSE, color = "black", size = 0.1,
+#'                fill = "white")
+#'
+#' # 50nm resolution, coastlines only
+#' tracking_chart(countries = FALSE, res = 50, color = "black", size = 0.1,
+#'                fill = "white")
+#'
+#' # Adding and modifying with ggplot functions
+#' tracking_chart(color = "black", size = 0.1, fill = "white") +
+#'     ggplot2::labs(x = "Lon", y = "Lat", title = "Base Tracking Chart")
+#' }
 #' @export
 tracking_chart <- function(countries = TRUE, states = TRUE, res = 110, ...) {
 
@@ -55,28 +108,30 @@ tracking_chart <- function(countries = TRUE, states = TRUE, res = 110, ...) {
     # countries is FALSE, return coastlines data. Otherwise, build countries w/
     # states if states is TRUE.
     if (!countries) {
-        x <- get(paste0("coastline", res))
+        dataset <- paste0("coastline", res)
+        base_map_data <- getExportedValue(ns = pkg, name = dataset)
     } else {
-        x <- get(paste0("countries", res))
+        dataset <- paste0("countries", res)
+        base_map_data <- getExportedValue(ns = pkg, name = dataset)
         if (states) {
             if (res >= 50) {
-                # Resolution for states is only 10 or 50. If res is > 50, use
-                # rnaturalearthdata::states50
-                y <- get(paste0("states", 50))
+                dataset <- paste0("states", 50)
+                state_map_data <- getExportedValue(ns = pkg, name = dataset)
             } else {
-                y <- get(paste0("states", res))
+                dataset <- paste0("states", res)
+                state_map_data <- getExportedValue(ns = pkg, name = dataset)
             }
         }
     }
 
     p <- ggplot2::ggplot() +
-        ggplot2::geom_polygon(data = x,
+        ggplot2::geom_polygon(data = base_map_data,
                               ggplot2::aes(long, lat, group = group), ...) +
         ggplot2::coord_equal()
 
-    if (exists("y"))
+    if (exists("state_map_data"))
         p <- p +
-        ggplot2::geom_polygon(data = y,
+        ggplot2::geom_polygon(data = state_map_data,
                               ggplot2::aes(long, lat, group = group), ...)
 
     return(p)