+++ R/update.R
@@ -25,20 +25,22 @@ create_df_update <- function() {
 #'   \item{Contents}{Text content of product}
 #' }
 #' @param link URL to storm's archive page.
-#' @param msg Show link being worked. Default, FALSE.
 #' @seealso \code{\link{get_storms}}, \code{\link{update}}
 #' @export
-get_update <- function(link, msg = FALSE) {
+get_update <- function(link) {
 
     # Check status of link(s)
     valid.link <- sapply(link, status)
     valid.link <- na.omit(valid.link)
-    if (length(valid.link) == 0)
-        stop("No valid links.")
 
     products <- purrr::map(valid.link, get_products) %>% purrr::flatten_chr()
 
-    products.update <- purrr::map(filter_update(products), update)
+    products <- filter_update(products)
+
+    # Set progress bar
+    p <- dplyr::progress_estimated(n = length(products))
+
+    products.update <- purrr::map(products, update, p)
 
     update <- purrr::map_df(products.update, dplyr::bind_rows)
 
@@ -50,13 +52,18 @@ get_update <- function(link, msg = FALSE) {
 #' @details Given a direct link to a cyclone update product, parse and return
 #' dataframe of values.
 #' @param link Link to a storm's specific update advisory product.
-#' @param msg Display each link as being worked; default is FALSE
+#' @param p dplyr::progress_estimate.
 #' @return Dataframe
 #' @seealso \code{\link{get_update}}
 #' @keywords internal
-update <- function(link, msg = FALSE) {
+update <- function(link, p) {
 
-    contents <- scrape_contents(link, msg = msg)
+    p$pause(0.5)$tick()$print()
+
+    contents <- scrape_contents(link)
+
+    # Replace all carriage returns with empty string.
+    contents <- stringr::str_replace_all(contents, "\r", "")
 
     # Make sure this is a update advisory product
     if (!any(stringr::str_count(contents, c("MIATCUAT", "MIATCUEP"))))
@@ -68,6 +75,10 @@ update <- function(link, msg = FALSE) {
     name <- scrape_header(contents, ret = "name")
     adv <- scrape_header(contents, ret = "adv")
     date <- scrape_header(contents, ret = "date")
+
+    if (getOption("rrricanes.working_msg"))
+        message(sprintf("Working %s %s Update #%s (%s)",
+                        status, name, adv, date))
 
     df <- df %>%
         tibble::add_row("Status" = status,