+++ tests/testthat/test_wndprb.R
@@ -1,18 +1,11 @@
 context("Wind Speed Probabilities (wndprb)")
 
-## ---- Data -------------------------------------------------------------------
-## ---- * Current Data ---------------------------------------------------------
-al2006 <- get_storms(year = 2006, basin = "AL") %>%
-    dplyr::select(Link) %>%
-    purrr::flatten_chr()
+## ---- Base Data --------------------------------------------------------------
+## ---- * 2008, AL -------------------------------------------------------------
+df.al092008.wndprb <- al2008 %>% dplyr::slice(9) %>% .$Link %>% get_wndprb()
+load(system.file("extdata", "al092008.wndprb.Rda", package = "rrricanes"))
 
-df.al012006.wndprb <- get_wndprb(al2006[1])
-
-## ---- * Saved Data -----------------------------------------------------------
-load(system.file("extdata", "al012006.wndprb.Rda", package = "Hurricanes"))
-
-## ---- Test wndprb ------------------------------------------------------------
-#' Test return of wndprb
-test_that("Test wndprb", {
-    expect_identical(al012006.wndprb, df.al012006.wndprb)
+## ---- Test get_wndprb() ------------------------------------------------------
+test_that("Test get_wndprb()", {
+    expect_identical(al092008.wndprb, df.al092008.wndprb)
 })